Make opentelemetry dependency optional
