+++ .drone.yml
@@ -22,10 +22,15 @@ steps:
     commands:
       - /root/.cargo/bin/cargo fmt -- --check
 
-  - name: check lemmy_api_common with minimal deps
+  - name: check with different features
     image: clux/muslrust:1.60.0
     commands:
+      # api with minimal deps
       - cargo check -p lemmy_api_common
+      # opentelemetry console
+      - cargo check --features console
+      # default features
+      - cargo check
 
   - name: cargo clippy
     image: clux/muslrust:1.60.0
