+++ Cargo.toml
@@ -14,7 +14,8 @@ doctest = false
 debug = 0
 
 [features]
-console = ["console-subscriber"]
+console = ["console-subscriber", "opentelemetry", "opentelemetry-otlp", "tracing-opentelemetry",
+    "reqwest-tracing/opentelemetry_0_16"]
 default = []
 
 [workspace]
@@ -54,14 +55,14 @@ tracing-actix-web = { version = "0.5.1", default-features = false }
 tracing-error = "0.2.0"
 tracing-log = "0.1.2"
 tracing-subscriber = { version = "0.3.9", features = ["env-filter"] }
-console-subscriber = { version = "0.1.3", optional = true }
 url = { version = "2.2.2", features = ["serde"] }
 reqwest = { version = "0.11.10", features = ["json"] }
 reqwest-middleware = "0.1.5"
-reqwest-tracing = { version = "0.2.1", features = ["opentelemetry_0_16"] }
+reqwest-tracing = "0.2.1"
 clokwerk = "0.3.5"
 doku = "0.11.0"
-opentelemetry = { version = "0.17.0", features = ["rt-tokio"] }
-opentelemetry-otlp = "0.10.0"
-tracing-opentelemetry = "0.17.2"
 parking_lot = "0.12.0"
+console-subscriber = { version = "0.1.3", optional = true }
+opentelemetry = { version = "0.17.0", features = ["rt-tokio"], optional = true }
+opentelemetry-otlp = { version = "0.10.0", optional = true }
+tracing-opentelemetry = { version = "0.17.2", optional = true }
